Cleaning FixedPrecision (#3253)
* Added Conv2d to nn

* Removed conv2d from precision.py

* Added Dropout to nn.functional

Removed dropout from precision.py

* Added tests for dropout in test_functional.py

Removed tests for dropout from test_precision.py

* Added conv2d to nn.functional

* Added tests for nn.functional.conv2d in test_functional.py

removed tests from test_precision.py and test_additive_shared.py

* Fix torch.nn.Conv2d and integrate with nn.functiona.conv2d

* Added tests for nn.Conv2d

* Commented tests for conv2d in test_additive_shared.py
